// web/src/domain/types/user.ts
export enum UserActionTypes {
  DISPATCH_SIGN_UP_ACTION = 'DISPATCH_SIGN_UP_ACTION',
  SIGN_UP_SUCCESS_ACTION = 'SIGN_UP_SUCCESS_ACTION',
  SIGN_UP_FAILURE_ACTION = 'SIGN_UP_FAILURE_ACTION',
  SIGN_UP_RESET_ACTION = 'SIGN_UP_RESET_ACTION',

  DISPATCH_DELETE_USER_ACTION = 'DISPATCH_DELETE_USER_ACTION',
  SET_DELETE_USER_ACTION = 'SET_DELETE_USER_ACTION',

  DISPATCH_UPDATE_USER_ACTION = 'DISPATCH_UPDATE_USER_ACTION',
  SET_UPDATE_USER_ACTION = 'SET_UPDATE_USER_ACTION',

  DISPATCH_GET_USERS_ACTION = 'DISPATCH_GET_USERS_ACTION',
  SET_GET_USERS_ACTION = 'SET_GET_USERS_ACTION',

  DISPATCH_RESET_PASSWORD_ACTION = 'DISPATCH_RESET_PASSWORD',
  SET_RESET_PASSWORD_ACTION = 'SET_RESET_PASSWORD_ACTION'
}

export enum UserSignUpStatus {
  IDLE = 'IDLE',
  SUCCESS = 'SUCCESS',
  PENDING = 'PENDING',
  FAILURE = 'FAILURE'
}